refactor(drive): split drive_document_count_query into 7 focused submodules
`drive_document_count_query/mod.rs` had grown to 2096 lines mixing
six unrelated concerns. Split each into its own file alongside
`tests.rs`. Pure refactor — zero behavior change.

## New layout

```
drive_document_count_query/
├── mod.rs                       ← types + module decls + re-exports (~150 lines)
├── tests.rs                     ← existing tests, unchanged          (1274 lines)
├── mode_detection.rs            ← operator classification + detect_mode (~190)
├── index_picker.rs              ← find_countable_*_index helpers     (~140)
├── path_query.rs                ← 3 path-query builders              (~450)
├── execute_point_lookup.rs      ← Equal/In execution paths           (~320)
├── execute_range_count.rs       ← range execution + RangeCountOptions (~300)
└── drive_dispatcher.rs          ← Drive::execute_document_count_*    (~680)
```

Net: -1989 / +43 in mod.rs, +2071 across the six new files. Tests
file is untouched.

## Seams

- **`mode_detection.rs`** — pure where-clause/flag shape classifier.
  No Drive, no contract, no indexes. Called from both server-side
  executors (to pick a dispatch arm) and the SDK verifier (to
  validate the request before reconstructing the path query).
- **`index_picker.rs`** — pure index lookup. Takes the document
  type's `BTreeMap<String, Index>` + where clauses, returns
  `Option<&Index>`. The two pickers are siblings because they share
  the same `is_indexable_for_count` / `is_range_operator`
  classification helpers from `mode_detection.rs`.
- **`path_query.rs`** — the load-bearing prover/verifier-agreement
  boundary. Three builders that produce the exact `PathQuery` bytes
  both prover and verifier must agree on. Pulling this into its own
  file makes the "if you touch this, prover and verifier must agree
  on the bytes" invariant much louder.
- **`execute_point_lookup.rs`** — Equal/In execution (no range).
  `execute_no_proof`, `execute_with_proof`, and the four private
  recursive helpers for the CountTree walk.
- **`execute_range_count.rs`** — range execution (3 methods) +
  `RangeCountOptions`. Uses `ProvableCountTree` boundary walks
  exclusively, shares nothing with the point-lookup half except the
  query struct.
- **`drive_dispatcher.rs`** — the only `impl Drive` blocks. Per-mode
  executors that pick an index for their mode and call the
  per-instance executor, plus `execute_document_count_request` (the
  abci entry point) and `DocumentCountRequest` /
  `DocumentCountResponse`.

## Cross-file visibility

The only private helper that's referenced across the split is
`is_indexable_for_count` (used by `detect_mode`,
`find_countable_index_for_where_clauses`, and
`find_range_countable_index_for_where_clauses`). Bumped from `fn`
to `pub(super) fn` in `mode_detection.rs` so the sibling submodules
can call it via `Self::*` method syntax. Everything else stays
private to its file (private helpers in
`execute_point_lookup.rs`, the executor recursion in
`expand_paths_and_count` / `count_recursive` / `fetch_count_at_path`,
the `range_clause_to_query_item` helper in `path_query.rs`).

## Feature gating

- `mode_detection.rs` and `index_picker.rs` — methods individually
  gated `#[cfg(any(feature = "server", feature = "verify"))]` for
  the ones that depend on `QuerySyntaxError` (which is gated by the
  same set). Module declared unconditionally so the always-compiled
  helpers (`is_indexable_for_count`, etc.) remain reachable.
- `path_query.rs` — same `feature = "server" || verify` gate at the
  file level (via `#![cfg]`), since every builder there depends on
  `QuerySyntaxError` and the grovedb types.
- `execute_point_lookup.rs`, `execute_range_count.rs`,
  `drive_dispatcher.rs` — `feature = "server"`-only via the
  parent's `#[cfg(feature = "server")] pub mod ...;` declaration in
  `mod.rs`. The file itself doesn't carry a redundant cfg (avoids
  clippy's `duplicated_attributes` warning).

`mod.rs` re-exports `WhereOperator` (from `super::conditions`) and
`QuerySyntaxError` so the existing `tests.rs` (which uses `use
super::*;`) still finds them.

Verified:
- `cargo check -p drive --features=verify --no-default-features` clean (verify-only build works).
- 33 `query::drive_document_count_query` unit tests pass.
- 27 `range_countable_index_e2e_tests` pass.
- 7 `drive-abci::query::document_count_query` tests pass.
- clippy clean on drive (server+verify), drive-abci, drive-proof-verifier, dash-sdk.
- fmt clean.
